+++ src/components/token/token.js
@@ -20,6 +20,23 @@ const shapeToClassMap = {
 
 export const SHAPES = Object.keys(shapeToClassMap);
 
+const colorToClassMap = {
+  tokenTint01: 'euiToken--tokenTint01',
+  tokenTint02: 'euiToken--tokenTint02',
+  tokenTint03: 'euiToken--tokenTint03',
+  tokenTint04: 'euiToken--tokenTint04',
+  tokenTint05: 'euiToken--tokenTint05',
+  tokenTint06: 'euiToken--tokenTint06',
+  tokenTint07: 'euiToken--tokenTint07',
+  tokenTint08: 'euiToken--tokenTint08',
+  tokenTint09: 'euiToken--tokenTint09',
+  tokenTint10: 'euiToken--tokenTint10',
+  tokenTint11: 'euiToken--tokenTint11',
+  tokenTint12: 'euiToken--tokenTint12',
+};
+
+export const COLORS = Object.keys(colorToClassMap);
+
 export const EuiToken = ({
   iconType,
   displayOptions,
@@ -52,8 +69,8 @@ export const EuiToken = ({
 
   const classes = classNames(
     'euiToken',
-    `euiToken--${tokenShape}`,
-    `euiToken--${tokenColor}`,
+    colorToClassMap[tokenColor],
+    shapeToClassMap[tokenShape],
     sizeToClassMap[size],
     {
       'euiToken--fill': fill,
@@ -89,8 +106,8 @@ EuiToken.propTypes = {
    * - `hideBorder`: disables the outer border
    */
   displayOptions: PropTypes.shape({
-    color: PropTypes.string,
-    shape: PropTypes.string,
+    color: PropTypes.oneOf(COLORS),
+    shape: PropTypes.oneOf(SHAPES),
     fill: PropTypes.boolean,
     hideBorder: PropTypes.boolean,
   }),